Trait statpack::cli::command::Command

pub trait Command<'a>: _Command<'a> + Parser {
    // Required methods
    fn create_option(&mut self, option: &str, description: &str) -> &mut Self;
    fn create_option_kwargs(
        &mut self,
        option: &str,
        description: &str,
        kwargs: &'a CommandOptionKwargs<'_>
    ) -> &mut Self;
    fn add_option(
        &mut self,
        short: &str,
        long: &str,
        description: &str
    ) -> &mut Self;
    fn add_option_kwargs(
        &mut self,
        short: &str,
        long: &str,
        description: &str,
        kwargs: &'a CommandOptionKwargs<'_>
    ) -> &mut Self;
}

The Command trait provides methods for managing command-line commands. It extends the _Command trait for internal apis and adds higher-level functionality for creating and managing options with additional configurations.

Required Methods

fn create_option(&mut self, option: &str, description: &str) -> &mut Self

Autogen a new command-line option and adds to option list

This method simplifies the process of building options by automatically generating the short, and long flags of the option based on the provided option string.

Example

let mut parser = Parser::new();
parser.create_option("verbose", "Enable verbose mode");

fn create_option_kwargs( &mut self, option: &str, description: &str, kwargs: &'a CommandOptionKwargs<'_> ) -> &mut Self

Similar to create_option but allows the use of predefined extra options

Example

let mut parser = Parser::new();
let kwargs = CommandOptionKwargsBuilder::new()
.set_deprecated()
.set_required();
parser.create_option("verbose", "Enable verbose mode", &kwargs);

fn add_option( &mut self, short: &str, long: &str, description: &str ) -> &mut Self

This function is useful for defining command-line arguments that the program accepts. By using this method, you can manually add options with a short and long version, along with a description to clarify the purpose of the option to the user.

Example

let mut parser = Parser::new();
parser.add_option("-v", "--verbose", "Enable verbose mode");

fn add_option_kwargs( &mut self, short: &str, long: &str, description: &str, kwargs: &'a CommandOptionKwargs<'_> ) -> &mut Self

Similar to add_option but allows the use of predefined extra options

Example

let mut parser = Parser::new();
let kwargs = CommandOptionKwargsBuilder::new()
.set_deprecated()
.set_required();
parser.add_option("-v", "--verbose", "Enable verbose mode", &kwargs);

Object Safety

This trait is not object safe.